Satisfy the -D warnings clippy gate in the shadow tests
`Comparison::Match` is a unit variant, so the `{ .. }` pattern is redundant
(`unneeded_struct_pattern`), and the ring-buffer test's loop counter is already
a `u64` (`useless_conversion`). Both only fire on test targets, which
`cargo clippy --all-targets -- -D warnings` lints in CI.

+++ autumn/src/shadow/registry.rs
@@ -348,7 +348,7 @@ mod tests {
             registry.record_comparison(
                 &context(),
                 diverging(&format!("{{\"a\":{n}}}"), "{}"),
-                1_000 + u64::from(n),
+                1_000 + n,
             );
         }
         let recent = registry.recent();
